ref(onboarding): Split python onboarding docs (#103029)
Contributes to
https://linear.app/getsentry/issue/TET-864/introduce-folders-for-onboarding-platforms
