+++ static/app/gettingStartedDocs/python/python/agentMonitoring.tsx
@@ -1,676 +1,15 @@
 import {ExternalLink} from 'sentry/components/core/link';
-import {SdkProviderEnum as FeatureFlagProviderEnum} from 'sentry/components/events/featureFlags/utils';
-import {
-  StepType,
-  type Docs,
-  type DocsParams,
-  type OnboardingConfig,
-  type OnboardingStep,
+import type {
+  OnboardingConfig,
+  OnboardingStep,
 } from 'sentry/components/onboarding/gettingStartedDoc/types';
-import {
-  getCrashReportBackendInstallSteps,
-  getCrashReportModalConfigDescription,
-  getCrashReportModalIntroduction,
-} from 'sentry/components/onboarding/gettingStartedDoc/utils/feedbackOnboarding';
+import {StepType} from 'sentry/components/onboarding/gettingStartedDoc/types';
 import {t, tct} from 'sentry/locale';
-import {
-  alternativeProfilingConfiguration,
-  getPythonInstallCodeBlock,
-  getPythonLogsOnboarding,
-  getPythonProfilingOnboarding,
-  getVerifyLogsContent,
-} from 'sentry/utils/gettingStartedDocs/python';
 
-type Params = DocsParams;
+import {getPythonInstallCodeBlock} from './utils';
 
-type FeatureFlagConfiguration = {
-  integrationName: string;
-  makeConfigureCode: (dsn: string) => string;
-  makeVerifyCode: () => string;
-  packageName: string;
-};
-
-const FEATURE_FLAG_CONFIGURATION_MAP: Record<
-  FeatureFlagProviderEnum,
-  FeatureFlagConfiguration
-> = {
-  [FeatureFlagProviderEnum.GENERIC]: {
-    integrationName: ``,
-    packageName: 'sentry-sdk',
-    makeConfigureCode: (dsn: string) => `sentry_sdk.init(
-    dsn="${dsn}",
-    # Add data like request headers and IP for users, if applicable;
-    # see https://docs.sentry.io/platforms/python/data-management/data-collected/ for more info
-    send_default_pii=True,
-    integrations=[
-        # your other integrations here
-    ],
-)`,
-    makeVerifyCode: () => `import sentry_sdk
-from sentry_sdk.feature_flags import add_feature_flag
-
-add_feature_flag('test-flag', False)  # Records an evaluation and its result.
-sentry_sdk.capture_exception(Exception("Something went wrong!"))`,
-  },
-
-  [FeatureFlagProviderEnum.LAUNCHDARKLY]: {
-    integrationName: `LaunchDarklyIntegration`,
-    packageName: "'sentry-sdk[launchdarkly]'",
-    makeConfigureCode: (dsn: string) => `import sentry_sdk
-from sentry_sdk.integrations.launchdarkly import LaunchDarklyIntegration
-import ldclient
-
-sentry_sdk.init(
-    dsn="${dsn}",
-    # Add data like request headers and IP for users, if applicable;
-    # see https://docs.sentry.io/platforms/python/data-management/data-collected/ for more info
-    send_default_pii=True,
-    integrations=[
-        LaunchDarklyIntegration(),
-    ],
-)`,
-    makeVerifyCode: () => `client = ldclient.get()
-client.variation("hello", Context.create("test-context"), False)  # Evaluate a flag with a default value.
-sentry_sdk.capture_exception(Exception("Something went wrong!"))`,
-  },
-
-  [FeatureFlagProviderEnum.OPENFEATURE]: {
-    integrationName: `OpenFeatureIntegration`,
-    packageName: "'sentry-sdk[openfeature]'",
-    makeConfigureCode: (dsn: string) => `import sentry_sdk
-from sentry_sdk.integrations.openfeature import OpenFeatureIntegration
-from openfeature import api
-
-sentry_sdk.init(
-    dsn="${dsn}",
-    # Add data like request headers and IP for users, if applicable;
-    # see https://docs.sentry.io/platforms/python/data-management/data-collected/ for more info
-    send_default_pii=True,
-    integrations=[
-        OpenFeatureIntegration(),
-    ],
-)`,
-    makeVerifyCode: () => `client = api.get_client()
-client.get_boolean_value("hello", default_value=False)  # Evaluate a flag with a default value.
-sentry_sdk.capture_exception(Exception("Something went wrong!"))`,
-  },
-
-  [FeatureFlagProviderEnum.STATSIG]: {
-    integrationName: `StatsigIntegration`,
-    packageName: "'sentry-sdk[statsig]'",
-    makeConfigureCode: (dsn: string) => `import sentry_sdk
-from sentry_sdk.integrations.statsig import StatsigIntegration
-from statsig.statsig_user import StatsigUser
-from statsig import statsig
-import time
-
-sentry_sdk.init(
-    dsn="${dsn}",
-    # Add data like request headers and IP for users, if applicable;
-    # see https://docs.sentry.io/platforms/python/data-management/data-collected/ for more info
-    send_default_pii=True,
-    integrations=[StatsigIntegration()],
-)
-statsig.initialize("server-secret-key")`,
-    makeVerifyCode: () => `while not statsig.is_initialized():
-    time.sleep(0.2)
-
-result = statsig.check_gate(StatsigUser("my-user-id"), "my-feature-gate")  # Evaluate a flag.
-sentry_sdk.capture_exception(Exception("Something went wrong!"))`,
-  },
-
-  [FeatureFlagProviderEnum.UNLEASH]: {
-    integrationName: `UnleashIntegration`,
-    packageName: "'sentry-sdk[unleash]'",
-    makeConfigureCode: (dsn: string) => `import sentry_sdk
-from sentry_sdk.integrations.unleash import UnleashIntegration
-from UnleashClient import UnleashClient
-
-sentry_sdk.init(
-    dsn="${dsn}",
-    # Add data like request headers and IP for users, if applicable;
-    # see https://docs.sentry.io/platforms/python/data-management/data-collected/ for more info
-    send_default_pii=True,
-    integrations=[UnleashIntegration()],
-)
-
-unleash = UnleashClient(...)  # See Unleash quickstart.
-unleash.initialize_client()`,
-    makeVerifyCode:
-      () => `test_flag_enabled = unleash.is_enabled("test-flag")  # Evaluate a flag.
-sentry_sdk.capture_exception(Exception("Something went wrong!"))`,
-  },
-};
-
-const getSdkSetupSnippet = (params: Params) => `
-import sentry_sdk
-
-sentry_sdk.init(
-    dsn="${params.dsn.public}",
-    # Add data like request headers and IP for users,
-    # see https://docs.sentry.io/platforms/python/data-management/data-collected/ for more info
-    send_default_pii=True,${
-      params.isLogsSelected
-        ? `
-    # Enable sending logs to Sentry
-    enable_logs=True,`
-        : ''
-    }${
-      params.isPerformanceSelected
-        ? `
-    # Set traces_sample_rate to 1.0 to capture 100%
-    # of transactions for tracing.
-    traces_sample_rate=1.0,`
-        : ''
-    }${
-      params.isProfilingSelected &&
-      params.profilingOptions?.defaultProfilingMode !== 'continuous'
-        ? `
-    # Set profiles_sample_rate to 1.0 to profile 100%
-    # of sampled transactions.
-    # We recommend adjusting this value in production.
-    profiles_sample_rate=1.0,`
-        : params.isProfilingSelected &&
-            params.profilingOptions?.defaultProfilingMode === 'continuous'
-          ? `
-    # Set profile_session_sample_rate to 1.0 to profile 100%
-    # of profile sessions.
-    profile_session_sample_rate=1.0,`
-          : ''
-    }
-)${
-  params.isProfilingSelected &&
-  params.profilingOptions?.defaultProfilingMode === 'continuous'
-    ? `
-
-def slow_function():
-    import time
-    time.sleep(0.1)
-    return "done"
-
-def fast_function():
-    import time
-    time.sleep(0.05)
-    return "done"
-
-# Manually call start_profiler and stop_profiler
-# to profile the code in between
-sentry_sdk.profiler.start_profiler()
-
-for i in range(0, 10):
-    slow_function()
-    fast_function()
-
-# Calls to stop_profiler are optional - if you don't stop the profiler, it will keep profiling
-# your application until the process exits or stop_profiler is called.
-sentry_sdk.profiler.stop_profiler()`
-    : ''
-}`;
-
-const onboarding: OnboardingConfig = {
-  install: () => [
-    {
-      type: StepType.INSTALL,
-      content: [
-        {
-          type: 'text',
-          text: tct('Install our Python SDK:', {
-            code: <code />,
-          }),
-        },
-        getPythonInstallCodeBlock(),
-      ],
-    },
-  ],
-  configure: (params: Params) => [
-    {
-      type: StepType.CONFIGURE,
-      content: [
-        {
-          type: 'text',
-          text: t(
-            "Import and initialize the Sentry SDK early in your application's setup:"
-          ),
-        },
-        {
-          type: 'code',
-          language: 'python',
-          code: getSdkSetupSnippet(params),
-        },
-        alternativeProfilingConfiguration(params),
-      ],
-    },
-  ],
-  verify: (params: Params) => [
-    {
-      type: StepType.VERIFY,
-      content: [
-        {
-          type: 'text',
-          text: t(
-            'You can verify your setup by intentionally causing an error that breaks your application:'
-          ),
-        },
-        {
-          type: 'code',
-          language: 'python',
-          code: 'division_by_zero = 1 / 0',
-        },
-        getVerifyLogsContent(params),
-      ],
-    },
-  ],
-  nextSteps: (params: Params) => {
-    const steps = [];
-
-    if (params.isLogsSelected) {
-      steps.push({
-        id: 'logs',
-        name: t('Logging Integrations'),
-        description: t(
-          'Add logging integrations to automatically capture logs from your application.'
-        ),
-        link: 'https://docs.sentry.io/platforms/python/logs/#integrations',
-      });
-    }
-
-    return steps;
-  },
-};
-
-export const crashReportOnboardingPython: OnboardingConfig = {
-  introduction: () => getCrashReportModalIntroduction(),
-  install: (params: Params) => getCrashReportBackendInstallSteps(params),
-  configure: () => [
-    {
-      type: StepType.CONFIGURE,
-      content: [
-        {
-          type: 'text',
-          text: getCrashReportModalConfigDescription({
-            link: 'https://docs.sentry.io/platforms/python/user-feedback/configuration/#crash-report-modal',
-          }),
-        },
-      ],
-    },
-  ],
-  verify: () => [],
-  nextSteps: () => [],
-};
-
-export const performanceOnboarding: OnboardingConfig = {
-  introduction: () =>
-    t(
-      "Adding Performance to your Python project is simple. Make sure you've got these basics down."
-    ),
-  install: onboarding.install,
-  configure: params => [
-    {
-      type: StepType.CONFIGURE,
-      content: [
-        {
-          type: 'text',
-          text: t(
-            "Configuration should happen as early as possible in your application's lifecycle."
-          ),
-        },
-        {
-          type: 'text',
-          text: tct(
-            "Once this is done, Sentry's Python SDK captures all unhandled exceptions and transactions. To enable tracing, use [code:traces_sample_rate=1.0] in the sentry_sdk.init() call.",
-            {code: <code />}
-          ),
-        },
-        {
-          type: 'code',
-          language: 'python',
-          code: `
-import sentry_sdk
-
-sentry_sdk.init(
-    dsn="${params.dsn.public}",
-    # Set traces_sample_rate to 1.0 to capture 100%
-    # of transactions for tracing.
-    traces_sample_rate=1.0,
-)`,
-        },
-        {
-          type: 'text',
-          text: tct(
-            'Learn more about tracing [linkTracingOptions:options], how to use the [linkTracesSampler:traces_sampler] function, or how to [linkSampleTransactions:sample transactions].',
-            {
-              linkTracingOptions: (
-                <ExternalLink href="https://docs.sentry.io/platforms/python/configuration/options/#tracing-options" />
-              ),
-              linkTracesSampler: (
-                <ExternalLink href="https://docs.sentry.io/platforms/python/configuration/sampling/" />
-              ),
-              linkSampleTransactions: (
-                <ExternalLink href="https://docs.sentry.io/platforms/python/configuration/sampling/" />
-              ),
-            }
-          ),
-        },
-      ],
-    },
-  ],
-  verify: () => [
-    {
-      type: StepType.VERIFY,
-      content: [
-        {
-          type: 'text',
-          text: tct(
-            'Verify that performance monitoring is working correctly with our [link:automatic instrumentation] by simply using your Python application.',
-            {
-              link: (
-                <ExternalLink href="https://docs.sentry.io/platforms/python/tracing/instrumentation/automatic-instrumentation/" />
-              ),
-            }
-          ),
-        },
-        {
-          type: 'text',
-          text: tct(
-            'You have the option to manually construct a transaction using [link:custom instrumentation].',
-            {
-              link: (
-                <ExternalLink href="https://docs.sentry.io/platforms/python/tracing/instrumentation/custom-instrumentation/" />
-              ),
-            }
-          ),
-        },
-      ],
-    },
-  ],
-  nextSteps: () => [],
-};
-
-export const featureFlagOnboarding: OnboardingConfig = {
-  install: () => [],
-  configure: ({featureFlagOptions = {integration: ''}, dsn}) => {
-    const {integrationName, packageName, makeConfigureCode, makeVerifyCode} =
-      FEATURE_FLAG_CONFIGURATION_MAP[
-        featureFlagOptions.integration as keyof typeof FEATURE_FLAG_CONFIGURATION_MAP
-      ];
-
-    return [
-      {
-        type: StepType.INSTALL,
-        content: [
-          {
-            type: 'text',
-            text:
-              featureFlagOptions.integration === FeatureFlagProviderEnum.GENERIC
-                ? t('Install the Sentry SDK.')
-                : t('Install the Sentry SDK with an extra.'),
-          },
-          getPythonInstallCodeBlock({packageName}),
-        ],
-      },
-      {
-        type: StepType.CONFIGURE,
-        content: [
-          {
-            type: 'text',
-            text:
-              featureFlagOptions.integration === FeatureFlagProviderEnum.GENERIC
-                ? `You don't need an integration for a generic usecase. Simply use this API after initializing Sentry.`
-                : tct('Add [name] to your integrations list.', {
-                    name: <code>{`${integrationName}`}</code>,
-                  }),
-          },
-          {
-            type: 'code',
-            language: 'python',
-            code: makeConfigureCode(dsn.public),
-          },
-        ],
-      },
-      {
-        type: StepType.VERIFY,
-        content: [
-          {
-            type: 'text',
-            text: t(
-              'Test your setup by evaluating a flag, then capturing an exception. Check the Feature Flags table in Issue Details to confirm that your error event has recorded the flag and its result.'
-            ),
-          },
-          {
-            type: 'code',
-            language: 'python',
-            code: makeVerifyCode(),
-          },
-        ],
-      },
-    ];
-  },
-  verify: () => [],
-  nextSteps: () => [],
-};
-
-export const mcpOnboarding: OnboardingConfig = {
-  install: () => {
-    const packageName = 'sentry-sdk';
-
-    return [
-      {
-        type: StepType.INSTALL,
-        content: [
-          {
-            type: 'text',
-            text: t(
-              'To enable MCP monitoring, you need to install the Sentry SDK with a minimum version of 2.43.0 or higher.'
-            ),
-          },
-          getPythonInstallCodeBlock({packageName}),
-        ],
-      },
-    ];
-  },
-  configure: (params: Params) => {
-    const mcpLowLevelStep: OnboardingStep = {
-      type: StepType.CONFIGURE,
-      content: [
-        {
-          type: 'text',
-          text: t('Configure Sentry for MCP low-level monitoring:'),
-        },
-        {
-          type: 'code',
-          language: 'python',
-          code: `
-import sentry_sdk
-from sentry_sdk.integrations.mcp import MCPIntegration
-
-sentry_sdk.init(
-    dsn="${params.dsn.public}",
-    traces_sample_rate=1.0,
-    # Add data like inputs and responses to/from MCP servers;
-    # see https://docs.sentry.io/platforms/python/data-management/data-collected/ for more info
-    send_default_pii=True,
-    integrations=[
-        MCPIntegration(),
-    ],
-)`,
-        },
-        {
-          type: 'text',
-          text: t('Set up your Low-level MCP server:'),
-        },
-        {
-          type: 'code',
-          language: 'python',
-          code: `
-from mcp.server.lowlevel import Server
-from mcp.types import Tool, TextContent
-
-server = Server("mcp-server")
-
-@server.list_tools()
-async def list_tools() -> list[Tool]:
-    """List all available tools."""
-    return [
-        Tool(
-            name="calculate_sum",
-            description="Add two numbers together",
-            inputSchema={
-                "type": "object",
-                "properties": {
-                    "a": {"type": "number", "description": "First number"},
-                    "b": {"type": "number", "description": "Second number"},
-                },
-                "required": ["a", "b"],
-            },
-        )
-    ]
-@server.call_tool()
-async def call_tool(name: str, arguments) -> list[TextContent]:
-    """Handle tool execution based on tool name."""
-
-    if name == "calculate_sum":
-        a = arguments.get("a", 0)
-        b = arguments.get("b", 0)
-        result = a + b
-        return [TextContent(type="text", text=f"The sum of {a} and {b} is {result}")]
-
-`,
-        },
-      ],
-    };
-
-    const mcpFastMcpStep: OnboardingStep = {
-      type: StepType.CONFIGURE,
-      content: [
-        {
-          type: 'text',
-          text: t('Configure Sentry for MCP low-level monitoring:'),
-        },
-        {
-          type: 'code',
-          language: 'python',
-          code: `
-import sentry_sdk
-from sentry_sdk.integrations.mcp import MCPIntegration
-
-sentry_sdk.init(
-    dsn="${params.dsn.public}",
-    traces_sample_rate=1.0,
-    # Add data like inputs and responses to/from MCP servers;
-    # see https://docs.sentry.io/platforms/python/data-management/data-collected/ for more info
-    send_default_pii=True,
-    integrations=[
-        MCPIntegration(),
-    ],
-)`,
-        },
-        {
-          type: 'text',
-          text: t('Set up your FastMCP server:'),
-        },
-        {
-          type: 'code',
-          language: 'python',
-          code: `
-from mcp.server.fastmcp import FastMCP
-# from fastmcp import FastMCP if you are using the standalone version
-
-mcp = FastMCP("mcp-server")
-
-@mcp.tool()
-async def calculate_sum(a: int, b: int) -> int:
-    """Add two numbers together."""
-    return a + b
-`,
-        },
-      ],
-    };
-
-    const manualStep: OnboardingStep = {
-      type: StepType.CONFIGURE,
-      content: [
-        {
-          type: 'text',
-          text: t('Configure Sentry for manual MCP instrumentation:'),
-        },
-        {
-          type: 'code',
-          language: 'python',
-          code: `
-import sentry_sdk
-
-sentry_sdk.init(
-    dsn="${params.dsn.public}",
-    traces_sample_rate=1.0,
-)
-`,
-        },
-      ],
-    };
-
-    const selected = (params.platformOptions as any)?.integration ?? 'mcp_fastmcp';
-    if (selected === 'mcp_fastmcp') {
-      return [mcpFastMcpStep];
-    }
-    if (selected === 'manual') {
-      return [manualStep];
-    }
-    return [mcpLowLevelStep];
-  },
-  verify: (params: Params) => {
-    const mcpVerifyStep: OnboardingStep = {
-      type: StepType.VERIFY,
-      content: [
-        {
-          type: 'text',
-          text: t(
-            'Verify that MCP monitoring is working correctly by triggering some MCP server interactions in your application.'
-          ),
-        },
-      ],
-    };
-
-    const manualVerifyStep: OnboardingStep = {
-      type: StepType.VERIFY,
-      content: [
-        {
-          type: 'text',
-          text: t(
-            'Verify that MCP monitoring is working correctly by running your manually instrumented code:'
-          ),
-        },
-        {
-          type: 'code',
-          language: 'python',
-          code: `
-import json
-import sentry_sdk
-
-# Invoke Agent span
-with sentry_sdk.start_span(op="mcp.server", name="tools/call calculate_sum") as span:
-    span.set_data("mcp.method.name", "tools/call")
-    span.set_data("mcp.request.argument.a", "1")
-    span.set_data("mcp.request.argument.b", "2")
-    span.set_data("mcp.request.id", 123)
-    span.set_data("mcp.session.id", "c6d1c6a4c35843d5bdf0f9d88d11c183")
-    span.set_data("mcp.tool.name", "calculate_sum")
-    span.set_data("mcp.tool.result.content_count", 1)
-    span.set_data("mcp.transport", "stdio")
-`,
-        },
-      ],
-    };
-    const selected = (params.platformOptions as any)?.integration ?? 'mcp_fastmcp';
-    if (selected === 'manual') {
-      return [manualVerifyStep];
-    }
-    return [mcpVerifyStep];
-  },
-  nextSteps: () => [],
-};
-
-export const agentMonitoringOnboarding: OnboardingConfig = {
-  install: (params: Params) => {
+export const agentMonitoring: OnboardingConfig = {
+  install: params => {
     const selected = (params.platformOptions as any)?.integration ?? 'openai_agents';
     let packageName = 'sentry-sdk';
 
@@ -699,7 +38,7 @@ export const agentMonitoringOnboarding: OnboardingConfig = {
       },
     ];
   },
-  configure: (params: Params) => {
+  configure: params => {
     const openaiAgentsStep: OnboardingStep = {
       type: StepType.CONFIGURE,
       content: [
@@ -1050,7 +389,7 @@ sentry_sdk.init(
     }
     return [openaiAgentsStep];
   },
-  verify: (params: Params) => {
+  verify: params => {
     const openaiAgentsVerifyStep: OnboardingStep = {
       type: StepType.VERIFY,
       content: [
@@ -1394,19 +733,5 @@ with sentry_sdk.start_span(op="gen_ai.chat", name="chat o3-mini") as span:
     }
     return [openaiAgentsVerifyStep];
   },
+  nextSteps: () => [],
 };
-
-const logsOnboarding = getPythonLogsOnboarding();
-
-const docs: Docs = {
-  onboarding,
-  performanceOnboarding,
-  crashReportOnboarding: crashReportOnboardingPython,
-  featureFlagOnboarding,
-  profilingOnboarding: getPythonProfilingOnboarding({traceLifecycle: 'manual'}),
-  agentMonitoringOnboarding,
-  mcpOnboarding,
-  logsOnboarding,
-};
-
-export default docs;
